backend/requirements.txt
Django
Djangorestframework
psycopg2
django-heroku
gunicorn